reuse status from queries
